backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
httpx
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
email-validator
aiosqlite
pytest
pytest-asyncio
httpx
python-multipart
pypdf
python-docx
